Split Bootstrap code from CoreRuntime
