+++ src/Umbraco.Core/Services/IRuntime.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using Microsoft.Extensions.DependencyInjection;
-using Umbraco.Core.Composing;
-using Umbraco.Core.Configuration;
 
 namespace Umbraco.Core
 {
@@ -11,16 +8,11 @@ namespace Umbraco.Core
     public interface IRuntime
     {
         /// <summary>
-        /// Boots the runtime.
-        /// </summary>
-        void Configure(IServiceCollection services);
-
-        /// <summary>
         /// Gets the runtime state.
         /// </summary>
         IRuntimeState State { get; }
 
-        void Start(IServiceProvider serviceProvider);
+        void Start();
 
         /// <summary>
         /// Terminates the runtime.
